Generate proposal particles.

Generate particles for a particular subject from a mix of population level (hierarchical) distribution, from the particles (containing individual level distribution) and/or from the conditional on accepted individual level particles, a more efficient prposal method.

gen_particles(
  num_particles,
  mu,
  sig2,
  particle,
  ...,
  mix_ratio = c(0.5, 0.5, 0),
  prop_mu = NULL,
  prop_sig2 = NULL,
  epsilon = 1
)

Arguments

num_particles
The total number of particles to generate using a combination of the three methods.

mu
A vector of means for the multivariate normal

sig2
A covariate matrix for the multivariate normal

particle
A particle (re proposals for latent variables)

mix_ratio
A vector of floats betwen 0 and 1 and summing to 1 which give the ratio of particles to generate from the population level parameters, the individual random effects and the conditional parameters repectively

epsilon
Reduce the variance for the individual level samples by this factor

Value

The new proposals

Details

This function is used in burnin, adaptation and sampling using various combinations of the arguments.

Examples

psamplers:::gen_particles(100, rep(0.2, 7), diag(rep(0.1, 7)), rep(0.3, 7))

#>                 [,1]         [,2]          [,3]         [,4]         [,5]
#>   [1,] -0.5707304270  0.198238204  3.965522e-01  0.563159637 -0.376109329
#>   [2,]  0.1106006875  0.024904881  3.989016e-01  0.853018209 -0.315764135
#>   [3,]  0.0349251486  0.183365816  3.717105e-01 -0.089055841  0.348043427
#>   [4,]  0.4333053579  0.797197695  1.691852e-01 -0.095940918  0.194956069
#>   [5,]  0.4957878119  0.255810599  2.770601e-01  0.713411236  0.235429553
#>   [6,]  0.1116974309  0.100879679  5.375124e-01  0.222146964 -0.002108541
#>   [7,]  0.3406571983  1.071339544  2.147145e-01  0.382687648  0.237376501
#>   [8,]  0.1230817892  0.134829507  2.060645e-01  0.209347931  0.373870735
#>   [9,]  0.0857718093  0.267469012  5.397380e-01 -0.010319371  0.552262685
#>  [10,] -0.1085910603  0.536800697  2.416379e-01  0.354517997 -0.337413457
#>  [11,]  0.6228978467  0.274849937  6.168810e-01  0.365674822  0.391870579
#>  [12,]  0.1714360038  0.808530796  6.105878e-01  0.436788590  0.375893577
#>  [13,] -0.6260926502  0.150765305  3.372080e-01  0.079216453  0.334139889
#>  [14,]  0.1879508067  0.353733787  7.290119e-01  0.087941161  0.499261478
#>  [15,]  0.0775523161 -0.048375614 -1.341695e-01 -0.051572289 -0.355383055
#>  [16,]  0.0302921552  0.271823906  5.094146e-01  0.133945506 -0.242532444
#>  [17,]  0.3798070639  0.872569169  3.343520e-01 -0.332616586  0.278867769
#>  [18,]  0.3421297642  0.640130246  3.348922e-01  0.234021046  0.207050214
#>  [19,]  0.0329482294  0.260762982 -1.624602e-01  0.467587086  0.225842016
#>  [20,]  0.3436754274 -0.070438776  1.092758e-01  0.483011716  0.221283533
#>  [21,]  0.7934031805  0.442369672  5.098895e-01  0.617983850 -0.154083654
#>  [22,]  0.6846954142  0.335708314  2.386125e-01 -0.159871119  0.023540122
#>  [23,]  0.2121746258  0.087302373  4.475570e-01  0.454377288 -0.400851989
#>  [24,]  0.2831889876 -0.366234862 -4.926933e-02 -0.158292911  0.314997039
#>  [25,]  0.1674574211 -0.108027851  6.018219e-01  0.503852125  0.443091041
#>  [26,] -0.2032962983  0.103354257  8.994229e-01 -0.129404465 -0.162562676
#>  [27,]  0.3752488473  0.830276806  1.512627e-01  1.010937120  0.535833617
#>  [28,]  0.0744571047 -0.060338034  1.694061e-02  0.757759175  0.242055809
#>  [29,]  0.5925218493  0.393560152  6.421809e-02  0.630215646  0.177592913
#>  [30,]  0.2207086943 -0.147379016 -2.285199e-04 -0.452584837  1.037665858
#>  [31,]  0.4486696808 -0.201771523  3.714402e-01  0.223750571  0.376617768
#>  [32,]  0.4976355092  0.092819066  1.761013e-01  0.212713745  0.239307448
#>  [33,]  0.3076516752  0.050518350  4.241274e-01 -0.283499199  0.275080486
#>  [34,] -0.2941117139  0.222469045 -2.238652e-03 -0.067274360  0.413531122
#>  [35,] -0.0854951545  0.616672361  5.479105e-01  0.580664815 -0.252607560
#>  [36,]  0.1753700235  0.339591862  2.407690e-01 -0.062536801  0.040749939
#>  [37,]  0.6603774745 -0.062027444  2.916347e-01  0.048193759  0.008736164
#>  [38,] -0.2532559295  0.196741805  1.328851e-01 -0.086609930 -0.464758982
#>  [39,]  0.1675540710  0.275881879  2.192579e-01 -0.488610003  0.162729350
#>  [40,]  0.7937947496  0.882658762  4.244314e-01  0.442541440  0.102534991
#>  [41,]  0.3781564232  0.301977999  3.159526e-01  0.557285247 -0.097727833
#>  [42,]  0.0786529555  0.144949055  1.298782e-01 -0.119241013  0.352018677
#>  [43,] -0.2477860791  0.477261335  3.973680e-01  0.867960728  0.087383572
#>  [44,]  0.5736470018  0.262737089  7.338074e-02  0.394845092  0.824283178
#>  [45,]  0.1292613040 -0.069816186  4.768048e-01 -0.275328115  0.424386286
#>  [46,]  0.2187543904 -0.034042804  3.628589e-01  0.505716772  0.408754946
#>  [47,]  0.9724408046 -0.166814450  4.289857e-01  0.740814485  0.413930001
#>  [48,] -0.3259902950  0.682125080  2.464689e-01  0.393296600  0.700452055
#>  [49,]  0.3347829622  0.569712372  2.258959e-01  0.959868864  0.264924215
#>  [50,]  0.8390446130  0.274678494  1.617195e-01  0.262300877  0.548701395
#>  [51,]  0.4664806342  0.848205533  7.581056e-01  0.780094828  0.178566244
#>  [52,]  0.0262246918  0.191270939  6.361110e-01  0.557111579  0.870311514
#>  [53,] -0.2417938607  0.036892324  6.480061e-01  0.245033256  0.356545325
#>  [54,] -0.0084558496  0.192932709  6.644041e-01  0.428104242  0.151081231
#>  [55,]  0.4052861358  0.190240990  2.688355e-01  0.310993993  0.422104086
#>  [56,]  0.5943587282  0.083463056  4.066957e-01  0.169670615  0.595439310
#>  [57,]  0.3026911971  0.943267323 -1.242784e-01  0.666501398  0.235739685
#>  [58,]  0.0327091714 -0.004075605  6.217589e-01 -0.173093733 -0.074672580
#>  [59,]  0.1059193797 -0.216918009  2.966982e-01  0.092394530  0.088274936
#>  [60,]  0.5012988199  0.462644725 -2.538334e-01  0.582580347  0.370530926
#>  [61,]  0.5343142467 -0.015419197 -6.293863e-01  0.527372716  0.079271425
#>  [62,]  0.5767968518  0.443512823  3.097969e-02  0.479048671  0.664361412
#>  [63,] -0.0687381255  0.148315359  3.364763e-01 -0.259105998 -0.145134511
#>  [64,]  0.1835246655  0.446122556  1.979112e-01 -0.107053064 -0.025727326
#>  [65,]  0.8627088234  0.422263433  9.483057e-03 -0.201011188  0.273418244
#>  [66,]  0.1874953877  0.662510783  2.300541e-01  0.621976477  0.216604636
#>  [67,]  0.0078251882  0.212868696  1.123505e-01  0.261329806  0.673080537
#>  [68,]  0.0445717745  0.027889192 -3.859735e-01 -0.135077709  0.144057820
#>  [69,]  0.4536845213 -0.121656774  2.171592e-01  0.506302663  0.536808734
#>  [70,]  0.0914783102  0.469903859  5.361599e-01  0.899668115 -0.351451563
#>  [71,]  0.1585372383 -0.146102803  1.398563e-01  0.214693817 -0.043156115
#>  [72,]  0.7312728495  0.074996312  5.093993e-01  0.392063460  0.362599816
#>  [73,]  0.5172429302  0.523025010  6.135214e-01  0.661843296 -0.091777818
#>  [74,]  0.1660806663  0.237264504  1.703863e-02  0.585955003  0.325187292
#>  [75,]  0.0689075150  0.239870711  4.672417e-01  0.473991848  0.473825450
#>  [76,] -0.5879708900  0.011518136  5.258999e-02  0.089991583  0.498056539
#>  [77,]  0.4478291646  0.528947308  4.941123e-01  0.104612603  0.369799343
#>  [78,]  0.2734919909  0.732506824  1.015878e-01  0.231495848  0.083790705
#>  [79,]  0.4973840193  0.080085385  5.219157e-02 -0.001404229  0.866919247
#>  [80,]  0.0451356321  0.333414112  1.945065e-01  0.819197916  0.096379089
#>  [81,] -0.2410093710  0.752532229 -5.364896e-01 -0.026491699  0.076279471
#>  [82,]  0.1999669830  0.034487972 -1.284659e-01  0.041462212  0.099385104
#>  [83,]  0.8714509663  0.017283916  2.853181e-01  0.151000074  0.133445166
#>  [84,]  0.3527750041  0.410835424  7.533761e-01  0.542200997  0.669202453
#>  [85,]  0.7805853981  0.196121811  2.830276e-01  0.121797864  0.064755331
#>  [86,]  0.3594513098  0.536807461 -3.060626e-01  0.374660062  0.498892533
#>  [87,]  0.1398982149 -0.051002971 -7.091356e-06  0.450791236  0.048535187
#>  [88,] -0.4759114213 -0.170945620  1.633429e-01  0.001938252  0.261593885
#>  [89,]  0.5107777657 -0.149528757  6.700042e-01 -0.144209742  0.648390690
#>  [90,] -0.0451532308  0.288283298  5.561615e-01  0.142089954  0.599793911
#>  [91,]  0.0008532751  0.391418156  5.773043e-01 -0.063315326 -0.059744741
#>  [92,]  0.5658848774  0.550305068  2.466053e-01  0.493759742  0.056138136
#>  [93,]  0.3712586114  0.069180544 -8.658436e-02  0.428643090  0.062509082
#>  [94,]  0.2085917395  0.333097102  5.277430e-01  0.106770262 -0.050036964
#>  [95,] -0.0865889417  0.622999882  4.470981e-01  0.549969417  0.258806653
#>  [96,]  0.3486587548  0.430651025  1.084505e+00 -0.389362851  0.433092227
#>  [97,]  0.2841437703  0.431848905  9.995825e-02  0.663708313  0.225535980
#>  [98,] -0.3484518653  0.061817807  2.575806e-01  0.618056463  0.985324640
#>  [99,]  0.3154364799  0.055907384  5.168687e-02  0.070227947  0.515628767
#> [100,]  0.1154090975 -0.339702984  4.275994e-01  0.473930188  0.308967309
#>                [,6]         [,7]
#>   [1,]  0.121788872  0.122777304
#>   [2,]  0.362043630 -0.389135962
#>   [3,]  0.314775265 -0.212532891
#>   [4,] -0.061453624 -0.278262763
#>   [5,]  0.157626424 -0.404022693
#>   [6,]  0.184199712  0.120473953
#>   [7,] -0.404539100  0.472615682
#>   [8,] -0.519138261  1.048299065
#>   [9,]  0.122240727 -0.172378215
#>  [10,] -0.265087657  0.289856228
#>  [11,]  0.165235288  0.254448639
#>  [12,]  0.026625830  0.551181968
#>  [13,]  0.536182369  0.531632049
#>  [14,]  0.616417057  0.106194188
#>  [15,] -0.018367256  0.023373513
#>  [16,]  0.281756669  0.060290746
#>  [17,]  0.539261244  0.844905186
#>  [18,]  0.390878561  0.116942596
#>  [19,] -0.212714236 -0.098807430
#>  [20,]  0.148557225 -0.061618447
#>  [21,]  0.362730751 -0.277219270
#>  [22,]  0.532841910  0.414302385
#>  [23,]  0.495920974  0.102269333
#>  [24,]  0.109594304  0.363701354
#>  [25,]  0.527590073  0.050143749
#>  [26,] -0.329785010  0.682544172
#>  [27,]  0.561351849  0.555389047
#>  [28,]  0.319059542  0.560090977
#>  [29,]  0.113937448 -0.573708160
#>  [30,] -0.164736595  0.092280845
#>  [31,]  0.331363038 -0.259257511
#>  [32,] -0.115732095  0.590032185
#>  [33,] -0.215148316  0.436231181
#>  [34,]  0.564729451 -0.333319627
#>  [35,]  0.637314812  0.200988509
#>  [36,] -0.177462484 -0.037715787
#>  [37,]  0.661727380  0.247332768
#>  [38,]  0.798733150 -0.106148271
#>  [39,]  0.235510730  0.202493835
#>  [40,]  0.520023084 -0.090629633
#>  [41,]  0.268886311  0.647592667
#>  [42,]  0.707358145 -0.279092822
#>  [43,] -0.136613139  0.540614157
#>  [44,]  0.795982556 -0.302365926
#>  [45,]  0.313323974  0.673042567
#>  [46,]  0.767236742  0.322776526
#>  [47,]  0.506986964  0.632720758
#>  [48,] -0.059027062 -0.057670616
#>  [49,] -0.140456913 -0.061224058
#>  [50,]  0.116921438  0.254011981
#>  [51,]  0.877787713  0.125602996
#>  [52,]  0.266774362  0.610679032
#>  [53,]  0.079137216 -0.003720692
#>  [54,]  0.257862328  0.687911021
#>  [55,]  0.306587413  0.302399150
#>  [56,]  0.881959252  0.077116452
#>  [57,]  0.180456875  0.849006517
#>  [58,]  0.499297608  0.964474388
#>  [59,]  0.148656861  0.717248723
#>  [60,]  0.483670334  0.243767929
#>  [61,] -0.298975184  0.324127941
#>  [62,]  0.060869393  0.145282930
#>  [63,]  0.524261920 -0.092388991
#>  [64,]  0.778797082  0.409700546
#>  [65,] -0.359357309  0.301128298
#>  [66,]  0.824477128  0.055212270
#>  [67,]  0.502713817  0.100906632
#>  [68,]  0.116065479  0.247117451
#>  [69,]  0.583517256  0.770958146
#>  [70,]  0.320410564  0.216153029
#>  [71,]  0.414524772  0.193851148
#>  [72,] -0.080601021  0.287408745
#>  [73,]  1.139552569  0.250378428
#>  [74,] -0.098076121  0.624350110
#>  [75,]  0.091434389  0.318158338
#>  [76,]  0.139594033  0.385495761
#>  [77,]  0.656681708  0.873464614
#>  [78,]  0.159372648  0.491788917
#>  [79,]  0.410356121  0.381935806
#>  [80,]  0.485632201  0.252438368
#>  [81,]  0.078463472  0.470094311
#>  [82,]  0.558025866  0.395752410
#>  [83,]  0.269915701 -0.489103639
#>  [84,]  0.256694686  0.137173677
#>  [85,]  0.265518002  0.122650226
#>  [86,]  0.432159703  0.925105917
#>  [87,]  0.374100679 -0.086572980
#>  [88,]  0.722621585  0.027931487
#>  [89,]  0.520722637  0.026668593
#>  [90,]  0.244898590 -0.049822524
#>  [91,] -0.154497324  0.143738112
#>  [92,]  0.581009161  0.304178784
#>  [93,]  0.248734249  0.411315376
#>  [94,]  0.469119661  0.532765129
#>  [95,]  0.005817947  0.184623982
#>  [96,]  0.186939617  0.095444429
#>  [97,] -0.219596178 -0.175411177
#>  [98,]  1.034456866 -0.022686209
#>  [99,]  0.227317171 -0.177694209
#> [100,]  0.582443579  0.180922697